bump to ROCm 7.2 for U24

+++ components/install_rccl.sh
@@ -3,39 +3,42 @@ set -ex
 
 source ${UTILS_DIR}/utilities.sh
 
-rccl_metadata=$(get_component_config "rccl")
-rccl_branch=$(jq -r '.branch' <<< $rccl_metadata)
-rccl_commit=$(jq -r '.commit' <<< $rccl_metadata)
-rccl_version=$(jq -r '.version' <<< $rccl_metadata)
-rccl_url=$(jq -r '.url' <<< $rccl_metadata)
-rccl_sha256=$(jq -r '.sha256' <<< $rccl_metadata)
-#the content of this tar ball is rccl but its name is misleading
-TARBALL=$(basename ${rccl_url})
-rccl_folder=rccl-$(basename $TARBALL .tar.gz)
+# On Ubuntu 24.04, RCCL comes from ROCm packages; build from source on other distros
+if [[ $DISTRIBUTION != "ubuntu24.04" ]]; then
+    rccl_metadata=$(get_component_config "rccl")
+    rccl_branch=$(jq -r '.branch' <<< $rccl_metadata)
+    rccl_commit=$(jq -r '.commit' <<< $rccl_metadata)
+    rccl_version=$(jq -r '.version' <<< $rccl_metadata)
+    rccl_url=$(jq -r '.url' <<< $rccl_metadata)
+    rccl_sha256=$(jq -r '.sha256' <<< $rccl_metadata)
+    #the content of this tar ball is rccl but its name is misleading
+    TARBALL=$(basename ${rccl_url})
+    rccl_folder=rccl-$(basename $TARBALL .tar.gz)
 
-# due to https://github.com/ROCm/rccl/issues/1877
-# we need to resort to doing a git clone instead of downloading the rccl tarball, by specifying a branch to clone
-if [[ $rccl_branch != "" && $rccl_branch != "null" ]]; then
-    git clone --branch ${rccl_branch} https://github.com/ROCm/rccl.git ${rccl_folder}
-    pushd ${rccl_folder}
-    git checkout ${rccl_commit}
+    # due to https://github.com/ROCm/rccl/issues/1877
+    # we need to resort to doing a git clone instead of downloading the rccl tarball, by specifying a branch to clone
+    if [[ $rccl_branch != "" && $rccl_branch != "null" ]]; then
+        git clone --branch ${rccl_branch} https://github.com/ROCm/rccl.git ${rccl_folder}
+        pushd ${rccl_folder}
+        git checkout ${rccl_commit}
+        popd
+    else
+        download_and_verify ${rccl_url} ${rccl_sha256}
+        tar -xzf ${TARBALL}
+    fi
+    mkdir ./${rccl_folder}/build
+    pushd ./${rccl_folder}/build
+
+    # aggressively crank up the number of compiler given that we have 2TB of memory to spare on MI300X
+    sed -i -E 's/(target_compile_options\(\s*rccl\s+PRIVATE[^)]*-parallel-jobs=)12/\196/' ../CMakeLists.txt
+
+    CXX=/opt/rocm/bin/hipcc CMAKE_POLICY_VERSION_MINIMUM=3.5 cmake -DCMAKE_PREFIX_PATH=/opt/rocm/ -DCMAKE_INSTALL_PREFIX=/opt/rccl ..
+    make -j$(nproc)
+    make install
     popd
-else
-    download_and_verify ${rccl_url} ${rccl_sha256}
-    tar -xzf ${TARBALL}
+    rm -rf ${TARBALL} ${rccl_folder}
+    write_component_version "RCCL" ${rccl_version}
 fi
-mkdir ./${rccl_folder}/build
-pushd ./${rccl_folder}/build
-
-# aggressively crank up the number of compiler given that we have 2TB of memory to spare on MI300X
-sed -i -E 's/(target_compile_options\(\s*rccl\s+PRIVATE[^)]*-parallel-jobs=)12/\196/' ../CMakeLists.txt
-
-CXX=/opt/rocm/bin/hipcc CMAKE_POLICY_VERSION_MINIMUM=3.5 cmake -DCMAKE_PREFIX_PATH=/opt/rocm/ -DCMAKE_INSTALL_PREFIX=/opt/rccl ..
-make -j$(nproc)
-make install
-popd
-rm -rf ${TARBALL} ${rccl_folder}
-write_component_version "RCCL" ${rccl_version}
 
 if [[ $DISTRIBUTION == *"ubuntu"* ]]; then
     sysctl kernel.numa_balancing=0
@@ -44,14 +47,21 @@ fi
 echo "kernel.numa_balancing=0" | tee -a /etc/sysctl.conf
 echo "vm.max_map_count=1048576" | tee -a /etc/sysctl.conf
 
+# TODO: switch to building from ROCm/rocm-systems
 git clone https://github.com/ROCmSoftwarePlatform/rccl-tests
 pushd ./rccl-tests
 
 source /opt/hpcx*/hpcx-init.sh
 hpcx_load
 
-RCCLLIB="/opt/rccl/lib/librccl.so"
-RCCLDIR="/opt/rccl"
+if [[ $DISTRIBUTION == "ubuntu24.04" ]]; then
+    # RCCL from ROCm packages
+    RCCLLIB="/opt/rocm/lib/librccl.so"
+    RCCLDIR="/opt/rocm"
+else
+    RCCLLIB="/opt/rccl/lib/librccl.so"
+    RCCLDIR="/opt/rccl"
+fi
 
 echo "gfx942" > target.lst
 echo "gfx90a" >> target.lst
